+++ tests/scan_batcher_test.cpp
@@ -560,7 +560,7 @@ INSTANTIATE_TEST_CASE_P(
                         {ChanField::RANGE2, 0x87288b444ddb9c9e},
                         {ChanField::REFLECTIVITY, 0x6912ca3fa04b0d1f},
                         {ChanField::REFLECTIVITY2, 0xf58aa5594d9749dc},
-                        {ChanField::NEAR_IR, 0x824f28547ce120fa},
+                        {ChanField::NEAR_IR, 0xc99384623c5d9feb},
                         {ChanField::RAW32_WORD1, 0x1d53499b615d3702},
                         {ChanField::RAW32_WORD2, 0x8b8d857f5ddec68d}}},
         // single return
@@ -651,7 +651,7 @@ static const Fields single_field_info{
 static const Fields fusa_info{
     {ChanField::RANGE, {UINT32, 0, 0x7fff, -3}},  // uint16 => uint32
     {ChanField::REFLECTIVITY, {UINT8, 2, 0xff, 0}},
-    {ChanField::NEAR_IR, {UINT8, 3, 0xff, 0}},
+    {ChanField::NEAR_IR, {UINT16, 3, 0xff, -4}},
     {ChanField::RANGE2, {UINT32, 4, 0x7fff, -3}},  // uint16 => uint32
     {ChanField::REFLECTIVITY2, {UINT8, 6, 0xff, 0}},
     {ChanField::RAW32_WORD1, {UINT32, 0, 0, 0}},
@@ -670,9 +670,6 @@ int add_profiles() {
 
 }  // namespace alternatives
 
-/**
- *
- */
 TEST_P(ScanBatcherSnapshotTest, extended_profile_comp_test) {
     static const int profiles_added = alternatives::add_profiles();
     (void)profiles_added;  // suppress stupid warning